Phase 6: Code cleanup
- Remove unused components: HomePage.jsx, HomePageNew.jsx, CityDashboard.jsx
- Extract satellite utils to src/utils/satellite.js (getGOESConfig, getAvailableSectors, generateFrameUrl, preloadImage)
- Update WeatherMap and MapWidgetPopup to use shared satellite utils
